Implemented Wander
If willWander in AIAgent is true, the AI will wander inside its room. You can set a delay between when the AI reaches its desired point and when the AI starts moving to a new point. When the player enters the room, it will stop wandering and instead use its behavior tree to determine its behavior.

I created the scene wander test to test the wander behavior, however line 30 of RoomTrigger.cs must be commented out for wander to work in the scene, as otherwise the error on line 30 means the AI will never stop wandering.